backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
python-multipart==0.0.11
pydantic==2.9.2
faster-whisper
google-genai
PyJWT

